+++ reactjs/src/core/extensions/page-rating/PageRating.tsx
@@ -3,13 +3,14 @@ import Icon from '../../../components/Icon';
 import Rating from 'react-rating';
 import {faStar} from '@fortawesome/free-solid-svg-icons';
 import {faStar as faStarEmpty} from '@fortawesome/free-regular-svg-icons';
-import {useOptions, useSettings} from '~/hooks/setting';
+import {useConfig, useOptions, useSettings} from '~/hooks/setting';
 import {useDispatch, useSelector} from 'react-redux';
 import {failureSnackbar} from '~/core/notifications/NotificationActions';
 import {ratePage} from './PageRatingActions';
 import {StoreProps} from '~/store/reducers';
 import {PageRatingReducerProps} from '~/core/extensions/page-rating/pageRatingReducers';
 import {manageReducer} from '~/helpers/action';
+import {fireEvent} from '~/helpers/events';
 
 /**
  * Renders page rating.
@@ -18,6 +19,7 @@ export default function PageRating() {
     const settings = useSettings();
     const options = useOptions();
     const dispatch = useDispatch();
+    const config = useConfig();
     const [value, setValue] = useState<number>(settings.rating.value);
     const [count, setCount] = useState<number>(settings.rating.count);
     const [hasRated, setHasRated] = useState<boolean>(settings.rating.hasRated);
@@ -28,6 +30,11 @@ export default function PageRating() {
         manageReducer({
             reducer: rate,
             onSuccess: response => {
+                fireEvent(config.events, 'onRated', {
+                    pageId: config.postId,
+                    value: response.value,
+                    totalCount: response.count,
+                });
                 setValue(response.value);
                 setCount(response.count);
                 setHasRated(true);
@@ -45,7 +52,7 @@ export default function PageRating() {
         if (hasRated) {
             dispatch(failureSnackbar(settings.i18.already_rated));
         } else {
-            dispatch(ratePage(settings.postId, rating));
+            dispatch(ratePage(config.postId, rating));
         }
     }
 
@@ -56,7 +63,10 @@ export default function PageRating() {
     return (
         <div itemScope itemType='http://schema.org/Article' className='anycomment anycomment-rating'>
             <div
-                className={'anycomment anycomment-rating__stars' + (hasRated ? ' anycomment-rating__stars-readonly' : '')}>
+                className={
+                    'anycomment anycomment-rating__stars' + (hasRated ? ' anycomment-rating__stars-readonly' : '')
+                }
+            >
                 <Rating
                     start={0}
                     stop={5}